# Web and Software Architecture/doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhoto Pizzari Federico Project
  description: |-
    This OpenAPI document describes the API for the WasaPhoto project
    of the student Federico Pizzari (1936451).
  version: "1"
...
paths:
...
  /image/{img}:
    parameters:
    - $ref: '#/components/parameters/image'
    delete:
      tags: ["Photos"]
      summary: Deletes a photo
      description: |-
        Given an image, delete it if it's yours
      operationId: deletePhoto
      responses:
        "204":
          $ref: '#/components/responses/ok'
        '401':
          $ref: '#/components/responses/noAuth'
        '403':
          $ref: '#/components/responses/forbiddenChange'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalError'
      security:
        - bearerAuth: []
    get:
      tags: ["Photos"]
      summary: Gets a photo
      description: |-
        Given an image id, get the photo assossiated with it
      operationId: getPhoto
      responses:
        "200":
          $ref: '#/components/responses/giveImage'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalError'
      security:
        - bearerAuth: []
  /user/{me}/follow/{followedName}:
    parameters:
    - $ref: '#/components/parameters/usernameMe'
    - $ref: '#/components/parameters/usernameFollow'
    put:
      tags: ["User Interaction"]
      summary: "Follow a user"
      description: |-
        The user calling the action will follow the user provided via the path
      operationId: followUser
      responses:
        '204':
          $ref: '#/components/responses/ok'
        '401':
          $ref: '#/components/responses/noAuth'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':  
          $ref: '#/components/responses/notFound'
        '406':
          $ref: '#/components/responses/notAcceptable'
        '500':
          $ref: '#/components/responses/internalError'
      security:
        - bearerAuth: []
    delete:
      tags: ["User Interaction"]
      summary: "Unfollow a user"
      description: |-
        The user calling the action will unfollow the user provided via the path
      operationId: unfollowUser
      responses:
        '204':
          $ref: '#/components/responses/ok'
        '401':
          $ref: '#/components/responses/noAuth'
        '404':  
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalError'
      security:
        - bearerAuth: []
...
  /user/{me}/ban/{bannedUser}:
    parameters:
    - $ref: '#/components/parameters/usernameMe'
    - $ref: '#/components/parameters/usernameBan'
    put:
      tags: ["User Interaction"]
      summary: "Ban a user"
      description: |-
        The user calling the action will ban the user provided via the path
      operationId: banUser
      responses:
        '204':
          $ref: '#/components/responses/ok'
        '401':
          $ref: '#/components/responses/noAuth'
        '404':  
          $ref: '#/components/responses/notFound'
        '406':
          $ref: '#/components/responses/notAcceptable'
        '500':
          $ref: '#/components/responses/internalError'
      security:
        - bearerAuth: []
    delete:
      tags: ["User Interaction"]
      summary: "Unban a user"
      description: |-
        The user calling the action will unban the user provided via the path
      operationId: unbanUser
      responses:
        '204':
          $ref: '#/components/responses/ok'
        '401':
          $ref: '#/components/responses/noAuth'
        '404':  
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalError'
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:            
      type: http
      description: The token provided in the response of the login action
      scheme: bearer
      bearerFormat: integer
  parameters:
    image:
      name: img
      in: path
      required: true
      description: "The id of the image of which we want to post the comment"
      schema:
        $ref: '#/components/schemas/image'
...
    usernameFollow:
      name: followedName
      in: path
      required: true
      description: "The username of the user we want to follow/unfollow"
      schema:
        $ref: '#/components/schemas/username'
    usernameBan:
      name: bannedUser
      in: path
      required: true
      description: "The username of the user we want to ban/unban"
      schema:
        $ref: '#/components/schemas/username'
    usernameMe:
      name: me
      in: path
      required: true
      description: "Your username"
      schema:
        $ref: '#/components/schemas/username'
...
  schemas:
...
    image:
      title: image
      description: "Represents the id of an image/post on WasaPhoto"
      type: integer
      example: 123
...
    photo:
      title: photo
      description: "Represents a photo"
      minLength: 1
      maxLength: 1000000
      type: string
      format: binary
...
    username:
      title: username
      description: "Represents the username"
      type: string
      pattern: ".*"
      example: Maria
      minLength: 3
      maxLength: 16
...
    systemMessage:
      title: systemMessage
      description: "Represents a system message"
      type: object
      properties:
        message:
          type: string
          description: "General description of the error"
          minLength: 1
          maxLength: 2048
          pattern: ".*"
          example: "Internal Server Error"
...
  responses:
...
    giveImage:
     description: Gives an image
     content:
        image/png:
          schema:
            $ref: "#/components/schemas/photo"
...
    notFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/systemMessage"
    notAcceptable:
      description: You can't execute this operation on yourself
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/systemMessage"
    internalError:
      description: An internal error occurred, please try again later
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/systemMessage"
...
    forbiddenChange:
      description: You are trying to change other users info. Authenticate with the right account and try again
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/systemMessage"
    forbidden:
      description: You are unable to access this resource, probably you are banned from this user. Try again later
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/systemMessage"
    noAuth:
      description: The token was not provided or the provided token is not valid for this user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/systemMessage"
...
    ok:
      description: Operation successfully executed
...